Misc fixes required for local packages using TypeScript to work properly.
The biggest issue was the compileStep.inputPath would be relative to the package,
instead of relative to the project root.  This meant that the files would
not be read properly.

Another issue was that the package's TS files were attempting to be compiled
alongside the project's TS files.  The problem with this is that when
we hand the package's compiled TS to Meteor, it is during the project
compilation phase and not during the package compilation phase, so the
package's compilation phase would be missing all of its TS files. The
solution is to introduce zzz.ts-compiler.ts files inside the package
directories as well, so we know when each of their phases is about to
end and can contribute each one's TS files to its phase.